Move the hook to the correct file
